further refactored secret functionalities to shared lib
